[dfs] Initial implementation of *INFO.
There seems to be a problem with the load and execution address
formatting.  The Beeb seems to print some of these as sign
extended (e.g. on the emulator I am using *INFO prints FF1900 for the
file W.SKETCH on the Welcome disk, rather than the value 031900 which
is what this version of the code prints).