[dfs] dfs.cc: Put internal code in anonymous namespace.  Implement --help.
Simplify the namespace setup in dfs.cc.  We now have some code in the
DFS namespace and some code in the anonymous namespace, but we no
longer have an anonymous namespace within DFS (it was unnecessary).
Also, ensure that the name(), description() and operator() methods
of CommandInterface implementations are public.

We also implement a "--help" option.