[basic] Improvements to the Mac dialect.
On the basis of examination of selected test programs, I found that
the Mac dialect uses all three extension codes 0xC6, 0xC7, 0xC8,
though some of the resulting tokens are different or only present in
ARM or Mac but not both.   Updated both the implementation and the
bbcbasic.5 document to match.   Simplified some of the code in
tokens.c and reduced the memory allocation.