[dfs] move exceptions out into a separate shared header.
In a larger project this would be a (perhaps minor) code smell,
because the universe of ways in which an API can fail is a property of
that API, and so the exceptions that an interface can raise should
probably be defined along with that interface.  However, the code in
this directory is simple enough that we can regard it as a single
module.

Another perspective is that these exceptions are ultimately presented
to the user as failure messages and keeping them together provides an
opportunity to ensure there is some user-interface consistency.