[dfs] adjust decompression buffer sizes to get better code coverage.
We use both a large stdio buffer (for fread) and a small stdio
buffer (for inflate).  The buffer for inflate is small so that our
test inputs exercise the buffer filling logic of
decompress_image_file.  The stdio buffer is large so that there is no
I/O performance downside resulting from this.

The code makes use of the fact that auto objects are destroyed in
reverse order of declaration, which is a somewhat subtle
point. However the code has explicit comments about this and we use a
capture in a lambda to ensure that the stdio buffer lives longer than
the FILE object which refers to it (or else we get a compilation
error).