[basic] empty input program should produce empty output.
The behaviour of the BBC Micro for this case is interesting.  If you
have a program in memory and then LOAD an empty file, the existing
program is not replaced (though admittedly I made this check with an
empty file made with BUILD, so the load and execution addresses would
not have been the same as for a BASIC program).