[dfs] StorageConfiguration no longer owns pointers to drives.
The intent of this change is to move responsibility for management of
the lifetime of image files to an AbstractImageFile object instead of
assuming that an AbstractDrive is the unit of ownership.

This will allow us to have a single open file for a DSD file, which
would present two AbstractDrive instances (one for each side).