[dfs] present image files as AbstractDrive instances via a view.
The view (FileView) is an adapter which takes acount of the fact that
there may be data stored in the file which is not relevant for some
particular emulated drive.  For example a DSD file would present two
AbstractDrive instances, one for each side.  Although this change
provides the groundwork for DSD drive support, is doesn't implement
DSD support.

FileView only caters to image files which store the sectors of an
emulated device in monotonically increasing order.

This change likely also has a performance regression, as the sectors
of the view are not cached.