[dfs] separate Catalog implementation from the FileSystem.
The rationale for this is to preprare for support alternative DFS
implementations, for example:

 - Opus DDOS, which has up to eight volumes per disk, each with a
   catalog
 - Duggan HDFS, which has a hierarchical directory structure.

The interfaces between the objects are still a bit rough, this is just
a relatively simple change to factor out the Catalog class.  The
Catalog class makes use of 1 or 2 CatalogFragment instances (1 for
Acorn DFS, 2 for Watford DFS).  The idea is that HDFS can use the
CatalogFragment implementation for its directories, and similarly for
Opus DDOS for volumes.