[dfs] more advanced format-autodetection
The basic idea is that we make a list of candidate geometries for the
disc images inside the image file we're working with.  We attempt to
locate and verify catalogs at the locations implied by each geometry
and disqualify those where the catalog is missing.

If more than one possibility matches, we pick the one with the
smallest capacity.

This algorithm might be confused by double-sided disc images in which
side 2 is not formatted.