[dfs] When inserting objects into ostreams, insert them as a single field.
This avoids problems when inserting composite objects like volume IDs.
These look like "2E" and the 2 and the E are formatted separately;
that looks strange if the field width is set to a non-zero value.

Also, minor bugfix in cat; after "Drive" show the drive we're actually
getting a catalogue for, not the one which is current (via --drive).