[dfs] bugfix: discs containing empty files are valid
Empty files have a start sector but zero length.  The start sector is
often the same as the start sector of some other file, so we should
disregard them when we are trying to figure out whether a catalog
contains overlapping files.

We added an empty test file to an image file, so update the test that
uses it to include it; add a second test file to the same image so
that we continue to exercise the code path for catalogs havng an odd
number of file names in the default directory.